🚚🚧 move json-editor into atom.io/react-devtools
