+++ packages/atom.io/react-devtools/src/StateEditor.tsx
@@ -4,7 +4,8 @@ import { useI, useO } from "atom.io/react"
 import type { FC } from "react"
 
 import { ElasticInput } from "~/packages/hamr/react-elastic-input/src"
-import { JsonEditor } from "~/packages/hamr/react-json-editor/src"
+
+import { JsonEditor } from "./json-editor"
 
 export const fallback = <T,>(fn: () => T, fallbackValue: T): T => {
 	try {
